feat: move buffer to its own folder
