+++ src/handlers/dkg_round_3.rs
@@ -27,7 +27,7 @@ use ironfish_frost::frost::keys::KeyPackage;
 use ironfish_frost::participant::{Secret};
 use ledger_device_sdk::io::{Comm, Event};
 use crate::accumulator::accumulate_data;
-use crate::buffer::{Buffer, BUFFER_SIZE};
+use crate::nvm::buffer::{Buffer, BUFFER_SIZE};
 use crate::context::TxContext;
 use crate::handlers::dkg_get_identity::compute_dkg_secret;
 use crate::utils::{zlog, zlog_stack};
